Move embeddings cli reader to cli namespace
