+++ GPTifier/src/command_embed.cpp
@@ -1,6 +1,7 @@
 #include "command_embed.hpp"
 
 #include "api.hpp"
+#include "cli.hpp"
 #include "configs.hpp"
 #include "datadir.hpp"
 #include "help_messages.hpp"
@@ -9,7 +10,6 @@
 
 #include <fmt/core.h>
 #include <fstream>
-#include <getopt.h>
 #include <iostream>
 #include <json.hpp>
 #include <stdexcept>
@@ -17,57 +17,6 @@
 
 namespace
 {
-
-struct ParamsEmbedding
-{
-    bool print_help = false;
-    std::string input;
-    std::string input_file;
-    std::string model;
-};
-
-ParamsEmbedding read_cli_embed(const int argc, char **argv)
-{
-    ParamsEmbedding params;
-
-    while (true)
-    {
-        static struct option long_options[] = {{"help", no_argument, 0, 'h'},
-                                               {"model", required_argument, 0, 'm'},
-                                               {"input", required_argument, 0, 'i'},
-                                               {"read-from-file", required_argument, 0, 'r'},
-                                               {0, 0, 0, 0}};
-
-        int option_index = 0;
-        int c = getopt_long(argc, argv, "hm:i:r:", long_options, &option_index);
-
-        if (c == -1)
-        {
-            break;
-        }
-
-        switch (c)
-        {
-        case 'h':
-            params.print_help = true;
-            break;
-        case 'm':
-            params.model = optarg;
-            break;
-        case 'i':
-            params.input = optarg;
-            break;
-        case 'r':
-            params.input_file = optarg;
-            break;
-        default:
-            std::cerr << "Try running with -h or --help for more information\n";
-            exit(EXIT_FAILURE);
-        }
-    }
-
-    return params;
-}
 
 nlohmann::json select_embedding_model(const std::string &model)
 {
@@ -90,7 +39,7 @@ nlohmann::json select_embedding_model(const std::string &model)
     throw std::runtime_error("No model provided via configuration file or command line");
 }
 
-std::string get_post_fields(const ParamsEmbedding &params)
+std::string get_post_fields(const cli::ParamsEmbedding &params)
 {
     nlohmann::json body = select_embedding_model(params.model);
 
@@ -134,7 +83,7 @@ void export_embedding(const std::string &response, const std::string &input)
 
 void command_embed(const int argc, char **argv)
 {
-    ParamsEmbedding params = read_cli_embed(argc, argv);
+    cli::ParamsEmbedding params = cli::get_opts_embed(argc, argv);
 
     if (params.print_help)
     {
